Bugfix in the baze-core/util/IdentityWriter causing failure in case of chunked upload of data when Content-Length (Long) converted to Int32 overflows to a positive number.

+++ blaze-core/src/main/scala/org/http4s/blazecore/util/IdentityWriter.scala
@@ -21,9 +21,9 @@ private[http4s] class IdentityWriter[F[_]](size: Long, out: TailStage[ByteBuffer
 
   private var bodyBytesWritten = 0L
 
-  private def willOverflow(count: Long) =
-    if (size < 0) false
-    else count + bodyBytesWritten > size
+  private def willOverflow(count: Int) =
+    if (size < 0L) false
+    else count.toLong + bodyBytesWritten > size
 
   def writeHeaders(headerWriter: StringWriter): Future[Unit] = {
     headers = Http1Writer.headersToByteBuffer(headerWriter.result)
